# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAPI Laravel API",
        "description": "API REST que integra la Star Wars API (SWAPI) con caché, reintentos y DTOs. Gestiona planetas y habitantes localmente con relaciones Eloquent.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API v1 – local"
        }
    ],
    "paths": {
        "/swapi/personajes": {
            "get": {
                "tags": [
                    "SWAPI – Personajes"
                ],
                "summary": "Listar personajes de Star Wars (paginado)",
                "operationId": "fdc5efdf3470fc67da2f46b4005357ec",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de personajes en español"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "502": {
                        "description": "Error en SWAPI externa"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/personajes/{id}": {
            "get": {
                "tags": [
                    "SWAPI – Personajes"
                ],
                "summary": "Obtener un personaje por ID",
                "operationId": "a09884a8226f67d0f8602989ff550996",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personaje mapeado al español con DTO"
                    },
                    "401": {
                        "description": "API Key inválida"
                    },
                    "404": {
                        "description": "Personaje no encontrado en SWAPI"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/planetas": {
            "get": {
                "tags": [
                    "SWAPI – Planetas"
                ],
                "summary": "Listar planetas de Star Wars (paginado)",
                "operationId": "ce3960f2ffb8b4c28b4dd4cba48c4603",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de planetas en español"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/planetas/{id}": {
            "get": {
                "tags": [
                    "SWAPI – Planetas"
                ],
                "summary": "Obtener un planeta de SWAPI por ID",
                "operationId": "c62711bca32cea5d5b898a9c9f3a2520",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Planeta mapeado al español"
                    },
                    "404": {
                        "description": "Planeta no encontrado"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/naves": {
            "get": {
                "tags": [
                    "SWAPI – Naves"
                ],
                "summary": "Listar naves estelares de Star Wars (paginado)",
                "operationId": "1eae6076b56b78fc4801817b765668e4",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de naves en español"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/naves/{id}": {
            "get": {
                "tags": [
                    "SWAPI – Naves"
                ],
                "summary": "Obtener una nave estelar por ID",
                "operationId": "368ff08743bae324ea44ef8723e2f6f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nave mapeada al español"
                    },
                    "404": {
                        "description": "Nave no encontrada"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/peliculas": {
            "get": {
                "tags": [
                    "SWAPI – Películas"
                ],
                "summary": "Listar todas las películas de Star Wars",
                "operationId": "aa8244e53a9d295d33c11704530d226e",
                "responses": {
                    "200": {
                        "description": "Lista de películas en español"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/peliculas/{id}": {
            "get": {
                "tags": [
                    "SWAPI – Películas"
                ],
                "summary": "Obtener una película por ID",
                "operationId": "825f5d7429264807d8dbd425ea9b6c8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Película mapeada al español con DTO"
                    },
                    "404": {
                        "description": "Película no encontrada"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/especies": {
            "get": {
                "tags": [
                    "SWAPI – Especies"
                ],
                "summary": "Listar especies de Star Wars (paginado)",
                "operationId": "168e553160db6af39337df755fcc7bb6",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de especies en español"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/especies/{id}": {
            "get": {
                "tags": [
                    "SWAPI – Especies"
                ],
                "summary": "Obtener una especie por ID",
                "operationId": "54cc594cdee5532f309cde583c1362dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especie mapeada al español con DTO"
                    },
                    "404": {
                        "description": "Especie no encontrada"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/vehiculos": {
            "get": {
                "tags": [
                    "SWAPI – Vehículos"
                ],
                "summary": "Listar vehículos de Star Wars (paginado)",
                "operationId": "eff931b616ea9f01dd13156e9a5219d2",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de vehículos en español"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/swapi/vehiculos/{id}": {
            "get": {
                "tags": [
                    "SWAPI – Vehículos"
                ],
                "summary": "Obtener un vehículo por ID",
                "operationId": "5ec25acaec07aa0dc35522bbcdcc768c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo mapeado al español con DTO"
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key requerida. Configura API_KEY en tu .env y envíala como header X-API-KEY.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}